WS-1789 - Optimizely `referrer` attribute setting
